@@ -1,28 +1,27 @@
 // tag::copyright[]
 /*******************************************************************************
- * Copyright (c) 2020 IBM Corporation and others.
+ * Copyright (c) 2020, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
+ * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
+ * http://www.eclipse.org/legal/epl-2.0/
  *
- * Contributors:
- *     IBM Corporation - Initial implementation
+ * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 // end::copyright[]
 package io.openliberty.guides.models;
 
 import java.util.Objects;
 
-import javax.json.bind.Jsonb;
-import javax.json.bind.JsonbBuilder;
+import jakarta.json.bind.Jsonb;
+import jakarta.json.bind.JsonbBuilder;
 
 import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.serialization.Serializer;
 
 public class SystemLoad {
 
-    private static final Jsonb jsonb = JsonbBuilder.create();
+    private static final Jsonb JSONB = JsonbBuilder.create();
 
     public String hostname;
     public Double loadAverage;
@@ -37,8 +36,12 @@ public class SystemLoad {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof SystemLoad)) return false;
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof SystemLoad)) {
+            return false;
+        }
         SystemLoad sl = (SystemLoad) o;
         return Objects.equals(hostname, sl.hostname)
                 && Objects.equals(loadAverage, sl.loadAverage);
@@ -51,22 +54,23 @@ public class SystemLoad {
 
     @Override
     public String toString() {
-        return "CpuLoadAverage: " + jsonb.toJson(this);
+        return "CpuLoadAverage: " + JSONB.toJson(this);
     }
 
     public static class SystemLoadSerializer implements Serializer<Object> {
         @Override
         public byte[] serialize(String topic, Object data) {
-            return jsonb.toJson(data).getBytes();
+            return JSONB.toJson(data).getBytes();
         }
     }
       
     public static class SystemLoadDeserializer implements Deserializer<SystemLoad> {
         @Override
         public SystemLoad deserialize(String topic, byte[] data) {
-            if (data == null)
+            if (data == null) {
                 return null;
-            return jsonb.fromJson(new String(data), SystemLoad.class);
+            }
+            return JSONB.fromJson(new String(data), SystemLoad.class);
         }
     }
 }
